+++ src/tests/mod.rs
@@ -1,7 +1,7 @@
+use super::*;
+use crate::node::OSCQueryNode;
 use once_cell::sync::Lazy;
 use std::{collections::HashSet, fs};
-
-use super::*;
 
 pub mod vrc_dependent;
